- Change some sizes to 'size_t' - Change some offsets to 'off_t' - Need to unlock static icons - Move unbuffered logs flag to Config - Too much time debugging keep-alive problems with Netscape Navigator-2.x

+++ src/pass.c
@@ -1,6 +1,6 @@
 
 /*
- * $Id: pass.c,v 1.52 1997/07/19 04:01:24 wessels Exp $
+ * $Id: pass.c,v 1.53 1997/07/21 07:20:59 wessels Exp $
  *
  * DEBUG: section 39    HTTP Passthrough
  * AUTHOR: Duane Wessels
@@ -43,7 +43,7 @@ typedef struct {
 	int offset;
 	char *buf;
     } client, server;
-    int *size_ptr;		/* pointer to size for logging */
+    size_t *size_ptr;		/* pointer to size for logging */
     int proxying;
 } PassStateData;
 
@@ -366,10 +366,7 @@ passConnectDone(int fd, int status, void *data)
 }
 
 void
-passStart(int fd,
-    const char *url,
-    request_t * request,
-    int *size_ptr)
+passStart(int fd, const char *url, request_t * request, size_t *size_ptr)
 {
     /* Create state structure. */
     PassStateData *passState = NULL;
